+++ src/components/features/Products/atoms/ReceiptPrice.tsx
